+++ rollup.config.js
@@ -6,42 +6,55 @@ import cleanup from 'rollup-plugin-cleanup';
 import copy from 'rollup-plugin-copy';
 import { external } from '@qqi/rollup-external';
 
-export default {
-  input: './index.ts',
-  output: [
-    {
-      format: 'es',
-      entryFileNames: '[name].mjs',
-      preserveModules: true,
-      sourcemap: false,
-      exports: 'named',
-      dir: 'dist/',
-    },
-    {
-      format: 'cjs',
-      entryFileNames: '[name].cjs',
-      preserveModules: true,
-      sourcemap: false,
-      exports: 'named',
-      dir: 'dist/',
-    },
-  ],
-  // 配置需要排除的包
-  external: external({
-    include: ['src/object/createConstructor'],
+const formatEs = 'es';
+const formatCjs = 'cjs';
+const formatUmd = 'umd';
+
+/**
+ * ## 构建打包
+ *
+ */
+function createConfig({ format, dir = undefined, transform = false }) {
+  return {
+    input: './src/index.ts',
+    output: Object.fromEntries(
+      [
+        ['format', format], // 打包模式
+        ['dir', dir ?? `dist/${format}`], // 打包的目录
+        ['preserveModules', true], // 是否保留源码目录结构
+        ['preserveModulesRoot', 'src'], // 是否保持 src 目录结构（当前模式下貌似没有作用）
+        ['sourcemap', false], // 打包关闭 source map
+        ['exports', 'named'], // 导出模式
+        format === formatUmd && ['name', 'aJsTools'], // 在 umd 模式下设定全局变量名
+      ].filter(Boolean),
+    ),
+    plugins: [
+      resolve(),
+      commonjs(),
+      json(),
+      typescript({}),
+      // 去除无用代码
+      transform && cleanup(),
+      copy({
+        targets: [
+          { src: 'README.md', dest: 'dist' },
+          { src: 'LICENSE', dest: 'dist' },
+        ],
+      }),
+    ].filter(Boolean),
+    external: external(), // 排除 dependencies 依赖
+  };
+}
+
+export default [
+  createConfig({
+    format: formatEs,
   }),
-  plugins: [
-    resolve(),
-    commonjs(),
-    json(),
-    typescript({}),
-    // 去除无用代码
-    cleanup(),
-    copy({
-      targets: [
-        { src: 'README.md', dest: 'dist' },
-        { src: 'LICENSE', dest: 'dist' },
-      ],
-    }),
-  ],
-};
+  createConfig({
+    format: formatCjs,
+  }),
+  // createConfig({
+  //   format: formatUmd,
+  //   transform: true,
+  // }),
+];