+++ src/index.ts
@@ -1,5 +1,29 @@
-export { getRandomInt, getRandomFloat } from './getRandomNumber';
-
 export { toLowerCamelCase, toSplitCase } from './className';
 
+export type { CreateConstructor } from './object/createConstructor';
+export { createConstructor, ObjectAssign } from './object/createConstructor';
+
+export { getRandomFloat, getRandomInt } from './getRandomNumber';
+
 export { getRandomString } from './getRandomString';
+
+export { throttle, debounce } from './performance';
+
+export type {
+  DebounceAndThrottleReturnType,
+  debounce_throttle_options,
+} from './performance';
+
+export { escapeRegExp, autoEscapedRegExp } from './regexp';
+
+export { isBrowser, isNode } from './isNode';
+
+export {
+  intersection,
+  enArr,
+  union,
+  difference,
+  symmetricDifference,
+} from './array';
+
+export { sleep } from './sleep';